refactor: Remove dev_ prefix from files
