+++ src/nasem_dairy/NASEM_equations/animal_equations.py
