+++ src/nasem_dairy/NASEM_equations/fecal_equations.py
